Remove colors
This is temporary, the'll come back in a different form.

The 'problem' with the current colors is that they're applied to every
annotation. I'd like them to only be applied to keywords that appear in
multiple cells of the type/value,original/generics table. The colors are
intended to help us see where information of the original types and
values end up in the generics representations.